+++ src/components/Form/ActionComponent.tsx
@@ -1,13 +1,10 @@
-import React, { useCallback, useContext, useEffect, useState } from "react";
-import { MapBrowserEvent } from "ol";
-import { toLonLat } from "ol/proj";
+import React, { useContext, useState } from "react";
 
 import { useQueryClient } from "@tanstack/react-query";
 
-import { reverseGeocode } from "../../requests/geoapify/input";
-import { addPointToLayer } from "../MapComponent/addPointToLayer";
 import MapContext from "../MapComponent/MapContext";
 
+import EnableAddPointOnClick from "./EnableAddPointOnClick";
 import FileUploader from "./FileUploader";
 import RouteDisplay from "./RouteDisplay";
 import RouteInputs from "./RouteInputs";
@@ -18,9 +15,7 @@ import "../components.css";
 
 export const ActionComponent = () => {
   // Get layers from context
-  const { map, startLayer, endLayer, stopsLayer, routeLayer } =
-    useContext(MapContext);
-
+  const { routeLayer } = useContext(MapContext);
   const { start, end, stops } = useGetRoutePoints();
 
   // We use setState to make the hook above refresh whenever a point changes
@@ -35,39 +30,6 @@ export const ActionComponent = () => {
     routeLayer.clear();
   };
 
-  // Add points via click
-  const addPointOnClick = useCallback(
-    (e: MapBrowserEvent) => {
-      const { coordinate } = e;
-      reverseGeocode(toLonLat(coordinate)).then((searchResult) => {
-        if (!searchResult) {
-          return;
-        }
-        if (!start) {
-          addPointToLayer(searchResult, startLayer);
-          updateRoutePoints();
-        }
-
-        if (start && !end) {
-          addPointToLayer(searchResult, endLayer);
-          updateRoutePoints();
-        }
-        if (start && end) {
-          addPointToLayer(searchResult, stopsLayer, false);
-          updateRoutePoints();
-        }
-      });
-    },
-    [endLayer, startLayer, stopsLayer]
-  );
-
-  const [clickActive, setClickActive] = useState(false);
-
-  useEffect(() => {
-    clickActive && map.on("singleclick", addPointOnClick);
-    return () => map.un("singleclick", addPointOnClick);
-  }, [map, addPointOnClick, clickActive]);
-
   const canCreateRoute = start && end && !!stops.length;
 
   const [isShowingForm, setIsShowingForm] = useState(true);
@@ -78,18 +40,11 @@ export const ActionComponent = () => {
       <div className="action-component-wrapper">
         {isShowingForm && (
           <>
-            <div>
-              Create your best driving route between multiple points
-              <RouteInputs updateRoute={updateRoutePoints} />
-              <FileUploader updateFunction={updateRoutePoints} />
-              <StopsList updateFunction={updateRoutePoints} />
-            </div>
-            <div
-              onClick={() => setClickActive(!clickActive)}
-              className="map-click-btn"
-            >
-              {clickActive ? "Disable" : "Enable"} adding points from map click
-            </div>
+            Create your best driving route between multiple points
+            <RouteInputs updateRoute={updateRoutePoints} />
+            <FileUploader updateFunction={updateRoutePoints} />
+            <StopsList updateFunction={updateRoutePoints} />
+            <EnableAddPointOnClick refreshLayerCallback={updateRoutePoints} />
           </>
         )}
         {canCreateRoute && (
